Attempt to have arrays wrapped around arbitrary memory blocks
This commit is a work in progress. The idea is to add the possibility to
build Yorick arrays on top of anything suitable to store the array
values. This implies to add two members to the array structure: one for
the address of the first array value, the other for a "destructor"
function to free the memory backing the storage of the array values.